% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_extractData.R
\name{extractData}
\alias{extractData}
\title{Extract Data from Copernicus Marine}
\usage{
extractData(
  cm,
  variables,
  lon,
  lat,
  startDate,
  endDate,
  depth,
  frequency,
  csv_path = NULL,
  data_output = NULL
)
}
\arguments{
\item{cm}{Python object from the Copernicus Marine service.}

\item{variables}{A character vector of variable names to extract.}

\item{lon}{Longitude value(s) (either a single value or a range of two values).}

\item{lat}{Latitude value(s) (either a single value or a range of two values).}

\item{startDate}{Start date of the data extraction (in "YYYY-MM-DD" format).}

\item{endDate}{End date of the data extraction (in "YYYY-MM-DD" format).}

\item{depth}{Depth value(s) (either a single value or a range of two values).}

\item{frequency}{Frequency of the data, either "daily" or "monthly".}

\item{csv_path}{Path to the CSV file containing variable and dataset information. If \code{NULL}, the CSV file included in the package will be used.}

\item{data_output}{Path where the extracted data will be saved. If \code{NULL}, the current working directory will be used.}
}
\value{
A list containing the result of the last downloaded dataset. Additionally, metadata about the extracted data is saved to a CSV file.
}
\description{
This function extracts data from the Copernicus Marine service, using the user's credentials, geographic boundaries, and time ranges.
It requires a \code{.csv} file containing variable and dataset information, which should be located in the package's data folder.
The function will download data for the specified variables, time range, location, and depth and save metadata to a CSV file.
}
\examples{
\dontrun{
result <- extractData(
  cm,
  variables = c("temperature", "salinity"),
  lon = c(-10, 10),
  lat = c(35, 50),
  startDate = "2020-01-01",
  endDate = "2020-01-31",
  depth = c(0, 100),
  frequency = "daily"
)
}
}
