% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_listVariables.R
\name{listVariables}
\alias{listVariables}
\title{List Variable Names and Dataset Information from Copernicus Marine}
\usage{
listVariables(csv_path = NULL, show = NULL)
}
\arguments{
\item{csv_path}{A character string specifying the path to the CSV file containing variable and dataset information.
If \code{NULL}, the default CSV file included in the package will be used.}

\item{show}{A character string specifying the type of information to display. If \code{NULL}, only the names of available variables will be shown.
If "completeList", a full list of dataset details will be shown.}
}
\description{
This function lists "user-friendly" variable names, which have been developed to substitute the original variable names from the Copernicus Marine service.
It also retrieves additional information about each variable, including metrics, dataset name, resolution, and time and depth coverage.
The function requires a \code{.csv} file containing variable and dataset information, which should be located in the package's data folder.
}
\examples{
\dontrun{
listVariables(show = "completeList")
}
}
